+++ src/routes/mongo/MessageRouteMongo.ts
@@ -6,6 +6,7 @@ import { RouteDecorators } from "@rapidrest/service-core";
 import { AuditLogEntryMongo, DomainMongo, FocusedInboxOverrideMongo, FolderMongo, KeyVaultMongo, MailboxMongo, MatterMongo, MessageMongo } from "../../mongo.js";
 import { BaseMessageRoute } from "../BaseMessageRoute.js";
 import { RecoverableRepoUtils } from "../../util/RecoverableRepoUtils.js";
+import { buildMessageLabelFilterMongo } from "../../util/MessageListUtils.js";
 const { Model } = RouteDecorators;
 
 @Model(MessageMongo)
@@ -18,4 +19,8 @@ export class MessageRouteMongo extends BaseMessageRoute<MessageMongo> {
     protected domainClass: any = DomainMongo;
     protected matterClass: any = MatterMongo;
     protected keyVaultClass: any = KeyVaultMongo;
+
+    protected buildLabelUidsFilter(labelUids: string[]): Record<string, any> {
+        return buildMessageLabelFilterMongo(labelUids);
+    }
 }